Move StatsSet into item search folder
